Analyze Minesweeper Scenarios
I find Minesweeper interesting because it combines pure logic with probability.  This script is designed to help improve intuition on the probability side by making it possible to fairly quickly compute the set of solutions and sum them up to get an idea of the relative likelihood that a mine will be present in a given square.